chore: replace MSVC-style long/long long with fixed-width integer types
MSVC's long is 32-bit and long long is 64-bit, but long is 64-bit on
LP64 platforms (Linux/macOS). Use uint32_t/int32_t/uint64_t/int64_t
instead so widths are explicit and consistent across compilers.

+++ CCPAssert.cpp
@@ -52,7 +52,7 @@ public:
 		return m_result;
 	}
 
-    static unsigned long __stdcall MessageBoxThreadProc( void* p )
+    static uint32_t __stdcall MessageBoxThreadProc( void* p )
 	{
 		CCPAssertDialogThread* pThis = static_cast<CCPAssertDialogThread*>( p );
